First RentalPrice controller method added

+++ Models/Car.cs
@@ -6,6 +6,6 @@ namespace CarRentalAPI.Models
         public string Type { get; set; }
         public double Price { get; set; }
         public double Discount { get; set; }
-        public double DiscountDays { get; set; }
+        public int DiscountDays { get; set; }
     }
 }
